requirements.txt
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.2.3
django-redis==5.3.0
djangorestframework==3.14.0
drf-yasg==1.21.6
executing==1.2.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
inflection==0.5.1
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2==2.9.6
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.6
